LayoutTests/imported/w3c/web-platform-tests/svg-aam/role/roles.tentative-expected.txt


SVG

FAIL svg assert_equals: <svg height="100" width="100" data-testname="svg" data-expectedrole="UNDEFINED in SVG-AAM, but possibly graphics-document" class="ex">
  <circle cx="50" cy="50" r="40" stroke="black" stroke-width="3" fill="red"></circle>
</svg> expected "UNDEFINED in SVG-AAM, but possibly graphics-document" but got "generic"
FAIL svg-text assert_equals: <svg height="130" width="500" data-testname="svg-text" data-expectedrole="UNDEFINED in SVG-AAM, but possibly graphics-document" class="ex">
   <defs>
      <linearGradient id="grad1">
         <stop offset="0%" stop-color="yellow"></stop>
         <stop offset="100%" stop-color="red"></stop>
      </linearGradient>
   </defs>
   <ellipse cx="100" cy="70" rx="85" ry="55" fill="url(#grad1)"></ellipse>
   <text fill="#ffffff" font-size="45" font-family="Verdana" x="50" y="86">SVG</text>
</svg> expected "UNDEFINED in SVG-AAM, but possibly graphics-document" but got "generic"
FAIL svg-symbol assert_equals: <svg data-testname="svg-symbol" data-expectedrole="UNDEFINED in SVG-AAM, but possibly graphics-document" class="ex">
  <symbol id="myDot" width="10" height="10" viewBox="0 0 2 2">
    <circle cx="1" cy="1" r="1"></circle>
  </symbol>
</svg> expected "UNDEFINED in SVG-AAM, but possibly graphics-document" but got "generic"
